Fix datatype

+++ src/SockPerf.cpp
@@ -3312,7 +3312,7 @@ int bringup(const int *p_daemonize)
 			log_dbg("Warm-up set in relation to number of active connections. Warm up time: %" PRIu32 " usec; first connection's first packet TTL: %d usec; following connections' first packet TTL: %d usec\n",
 				g_pApp->m_const_params.warmup_msec,
 				TEST_FIRST_CONNECTION_FIRST_PACKET_TTL_THRESHOLD_MSEC * 1000,
-				TEST_ANY_CONNECTION_FIRST_PACKET_TTL_THRESHOLD_MSEC * 1000);
+				(int)(TEST_ANY_CONNECTION_FIRST_PACKET_TTL_THRESHOLD_MSEC * 1000));
 		}
 
 		os_set_signal_action(SIGINT, s_user_params.mode ? server_sig_handler : client_sig_handler);
